+++ apps/backoffice-v2/src/common/components/molecules/Pagination/Pagination.First.tsx
@@ -28,4 +28,5 @@ export const PaginationFirst: FunctionComponent<
     </span>
   </PaginationLink>
 );
+
 PaginationFirst.displayName = 'PaginationFirst';