+++ apps/backoffice-v2/src/common/components/molecules/Pagination/Pagination.Previous.tsx
@@ -28,4 +28,5 @@ export const PaginationPrevious: FunctionComponent<
     </span>
   </PaginationLink>
 );
+
 PaginationPrevious.displayName = 'PaginationPrevious';