+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/utils/get-tabs-block-map.tsx
@@ -51,6 +51,8 @@ export const getTabsToBlocksMap = ({
     documentReviewBlocks,
     businessInformationBlocks,
     caseOverviewBlock,
+    customDataBlock,
+    amlWithContainerBlock,
   ] = blocks;
 
   const defaultTabsMap = {
@@ -79,6 +81,7 @@ export const getTabsToBlocksMap = ({
     [Tab.UBOS]: [
       ...ubosUserProvidedBlock,
       ...ubosRegistryProvidedBlock,
+      ...amlWithContainerBlock,
       ...(createKycBlocks(blocksCreationParams?.workflow as TWorkflowById) || []),
     ],
     [Tab.ASSOCIATED_COMPANIES]: [
@@ -92,6 +95,7 @@ export const getTabsToBlocksMap = ({
       ...directorsDocumentsBlocks,
     ],
     [Tab.MONITORING_REPORTS]: [...websiteMonitoringBlocks],
+    [Tab.CUSTOM_DATA]: [...customDataBlock],
   } as const;
 
   if (theme?.type === WorkflowDefinitionConfigThemeEnum.KYB) {
@@ -106,7 +110,11 @@ export const getTabsToBlocksMap = ({
 
   if (theme?.type === WorkflowDefinitionConfigThemeEnum.KYC) {
     return {
-      [Tab.KYC]: [...businessInformationBlocks, ...createKycBlocks(blocksCreationParams?.workflow)],
+      [Tab.KYC]: [
+        ...businessInformationBlocks,
+        ...amlWithContainerBlock,
+        ...createKycBlocks(blocksCreationParams?.workflow),
+      ],
     } as const;
   }