+++ packages/common/src/rule-engine/operators/constants.ts
@@ -1,14 +1,15 @@
 import {
-  EQUALS,
-  NOT_EQUALS,
+  AML_CHECK,
   BETWEEN,
-  LT,
-  LTE,
+  EQUALS,
   GT,
   GTE,
-  LAST_YEAR,
-  EXISTS,
   IN,
+  LAST_YEAR,
+  LT,
+  LTE,
+  NOT_EQUALS,
+  EXISTS,
   NOT_IN,
 } from './helpers';
 
@@ -17,15 +18,14 @@ import { OPERATION } from './enums';
 export const OperationHelpers = {
   [OPERATION.EQUALS]: EQUALS,
   [OPERATION.NOT_EQUALS]: NOT_EQUALS,
+  [OPERATION.EXISTS]: EXISTS,
   [OPERATION.BETWEEN]: BETWEEN,
   [OPERATION.GT]: GT,
   [OPERATION.GTE]: GTE,
   [OPERATION.LT]: LT,
   [OPERATION.LTE]: LTE,
   [OPERATION.LAST_YEAR]: LAST_YEAR,
-  [OPERATION.EXISTS]: EXISTS,
   [OPERATION.IN]: IN,
   [OPERATION.NOT_IN]: NOT_IN,
+  [OPERATION.AML_CHECK]: AML_CHECK,
 } as const;
-
-export type OperationHelper = (typeof OperationHelpers)[keyof typeof OperationHelpers];