+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/ui
 
+## 0.5.11
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.16
+
 ## 0.5.10
 
 ### Patch Changes