+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.25
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.25
+
+## 0.6.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.24
+
 ## 0.6.23
 
 ### Patch Changes